omx/mm-core: Changing TARGET_PRODUCT to QCOM_TARGET_PRODUCT
Using TARGET_PRODUCT to refer to qcom boardnames conflicts with
the customers' definition of the variable (who use it to refer
to the end user visible name for the product). The variable
QCOM_TARGET_PRODUCT has been introduced to refer to the qcom
boardnames. This should be used instead.

Targets tested on: qsd8250_surf
Tests: make libmm-omxcore from android root

Change-Id: I16e1a379cfdad6dd79a693a15069d00232d6951d